gdn: allow the recurrent state cache in bfloat16
Behind GDN_BF16_RECURRENT_STATE, off by default, the linear attention
recurrent state cache is allocated in bfloat16 rather than float32. The
kernel widens the state to float32 on load and rounds it back on
writeback, so the recurrence is still computed in float32 and only the
stored checkpoint is narrower. The setting is carried to Ray workers,
because the worker process is the one that allocates the caches.

What it buys today is bandwidth, not capacity: the kernel moves half the
bytes per call, while the arithmetic that decides how many cache blocks
to allocate still reserves the float32 cache, so the freed memory does
not turn into more sequences until that arithmetic follows.

The option names the recurrent state of a gated delta net cache, which
holds a convolution state and then a recurrent one, and it refuses any
other cache rather than guessing. Both the kind of cache and the state
count are asked: Mamba1 and Mamba2 report two states whose second one is
an SSM state this makes no promise about, and the KDA cache reports a
gated delta net but carries four states whose second one is a second
convolution state. With the setting unset, every model allocates exactly
as it does today.

+++ tpu_inference/kernels/gdn/v3/vmem_ldst.py
@@ -202,7 +202,10 @@ def load_and_select_states(
             prev_conv_state = jnp.where(is_first_tile, prev_conv_state,
                                         prev_tile_conv)
 
-        hbm_recurrent_state = recurrent_slot_ref[idx, 0]
+        # Widen the recurrent state to float32 for the recurrence, which is
+        # an identity where the cache is already float32. The writeback
+        # rounds it back to whatever the cache holds.
+        hbm_recurrent_state = recurrent_slot_ref[idx, 0].astype(jnp.float32)
         prev_recurrent_state = jnp.where(has_initial_state,
                                          hbm_recurrent_state, 0)
